Misc updates and fixes
* Imagine: Combine IO::read/write and IO::readAtPos/writeAtPos into a single API with std::optional
* Imagine: Simplify IOUtils::getImpl() with std::optional
* Imagine: Rename single object overload of IOUtils::write() to put() for clarity
* Imagine: Implement sync() for MapIO
* Imagine: Keep track of directory depth in FSPicker for more logical back button handling
* EmuFramework: Revert autosave loading failure as fatal error when launching system
* EmuFramework: Don't implicitly round EmuSystem::audioMixRate()
* EmuFramework: Update each system's frameTime() so the mix rate is only applied when changed
* 2600.emu: Default to split console buttons in virtual controls
* GBC.emu: Flush backup memory before performing reset
* GBC.emu: Fix RTC detection from last commit and optimize saving/loading
* NES.emu: Update mappers: 134, 176, 205

+++ 2600.emu/src/main/MainSystem.hh
@@ -63,7 +63,7 @@ class A2600System final: public EmuSystem
 {
 public:
 	OSystem osystem;
-	double configuredInputVideoFrameRate{};
+	float configuredInputVideoFrameRate{};
 	Properties defaultGameProps{};
 	bool p1DiffB = true, p2DiffB = true, vcsColor = true;
 	Controller::Type autoDetectedInput1{};
